Add a Datasets class and tests
Also move other classes to submodules.
