+++ mapbox/service.py
@@ -1,0 +1,16 @@
+import os
+
+import requests
+
+
+class Service:
+    """Base service class"""
+
+    def get_session(self, token=None, env=None):
+        access_token = (
+            token or
+            (env or os.environ).get('MapboxAccessToken') or
+            (env or os.environ).get('MAPBOX_ACCESS_TOKEN'))
+        session = requests.Session()
+        session.params.update(access_token=access_token)
+        return session
