+++ tests/test_service.py
@@ -1,0 +1,33 @@
+import json
+import responses
+
+from mapbox.service import Service
+
+
+def test_service_session():
+    """Get a session using a token"""
+    session = Service().get_session('pk.test')
+    assert session.params.get('access_token') == 'pk.test'
+
+
+def test_service_session_env():
+    """Get a session using the env's token"""
+    session = Service().get_session(
+        env={'MapboxAccessToken': 'pk.test_env'})
+    assert session.params.get('access_token') == 'pk.test_env'
+
+
+def test_service_session_os_environ(monkeypatch):
+    """Get a session using os.environ's token"""
+    monkeypatch.setenv('MapboxAccessToken', 'pk.test_os_environ')
+    session = Service().get_session()
+    assert session.params.get('access_token') == 'pk.test_os_environ'
+    monkeypatch.undo()
+
+
+def test_service_session_os_environ_caps(monkeypatch):
+    """Get a session using os.environ's token"""
+    monkeypatch.setenv('MAPBOX_ACCESS_TOKEN', 'pk.test_os_environ')
+    session = Service().get_session()
+    assert session.params.get('access_token') == 'pk.test_os_environ'
+    monkeypatch.undo()
